Do not restrict email and phone logins if the respective providers' signups are disabled
Hi, this is related to supabase/supabase#4746. I think this might be the fix so I've decided to open a PR for discussion. I do not have a clear enough picture of the role of this condition other than what it was currently causing as per the linked issue.

I _think_ it should be okay to remove these conditions as I do not suppose preventing logins while the email and phone provider's signups are disabled is an intentional feature? If I am mistaken, please excuse the assumption. Should that be the case, could someone please explain the logic behind the decision? I am sure there might be a valid security reason I am overlooking. And if that's the case, is there an alternative for getting email signins to work while keeping signups disabled?

Thank you