fix(oauth): Remove password, phone on subsequent OAuth login
An unconfirmed user account using email + password login (or phone),
that is then connected to an OAuth account presents a security risk
since the user logging in with OAuth does not expect for there to be a
(potentially malicious) email + password or phone number. Therefore, the
password and phone identities are removed.