fix: properly escape `redirectTo` URL for magic links (#750)
If the user passes a proper URL with query params or a fragment in the
`redirectTo` parameter on sign-in with magic links, then the query
parameters and the fragment will "break" the URL because they are not
query escaped. Since some users have already noticed this and started
query escaping the `redirectTo` parameter, to keep backward
compatibility with them, the URL will only be escaped if it contains any
of the unescaped characters that should not appear within a query
parameter (i.e. `&=#`).