feat: no email password resets for users with no email identity (#793)
## What kind of change does this PR introduce?
* No longer allow password resets for user's without an email identity.
* Previously, it was possible to request for a password reset if you
have an oauth identity. This is because the oauth identity is always
guaranteed to have an email and the password reset would be allowed.
This means that the user will now be able to login with email + password
if the reset was successful even though no email identity is created for
them. The password reset acts as a backdoor access to the user's account
in the event the user loses access to their oauth account. However, the
user should attempt to regain their access to the account by liaising
with the oauth provider rather than initiate a password recovery to gain
access to the account.