feat: add `is_sso_user` column to `users` which allows duplicate emails to exist on those rows (#828)
Adds a new `duplicated_emails` boolean column on `auth.users`. This
column in combination with a partial unique index will allow rows in the
table to contain duplicate emails.

This is a somewhat temporary measure that ensures database consistency
until full programmatic support is available for account merging /
linking / delinking through APIs and hooks.

Cases such as SAML, which do not obey the email-similarity account
linking rule, may create distinct user accounts that may have the same
email address. In fact, this is the only time the `duplicated_emails`
column is set to `true`.