feat: remove `id_token` flow with freeform provider (#927)
The `POST /token?grant_type=id_token` implementation allowed any OpenID
Connect compliant issuer to be used for authentication, albeit in a very
limited and broken way. Given we are not supporting this at this time,
the implementation is being removed.

Existing OIDC providers, importantly `google` and `apple` continue to be
supported.