chore: amend email and phone enabled checks on resend  (#1221)
## What kind of change does this PR introduce?

Move the check for whether email logins are enabled into the resend
function. Also checks if Phone Logins are enabled before proceeding to
resend

## What is the current behavior?

Currently, the check is done in middleware. This would mean that resend
requests with phone based logins will not go through if email logins are
disabled.

---------

Co-authored-by: joel@joellee.org <joel@joellee.org>