feat: add cleanup for session timebox and inactivity timeout (#1298)
Builds on top of #1288.

Adds cleanup for timeboxed sessions and sessions that have expired due
to inactivity timeout.

It achieves backward compatibility with sessions that have `null` in
`refreshed_at` by looking at the `updated_at` column of the refresh
tokens table. This approach is the one that puts the least strain on the
database, having considered backfilling (very expensive at least
`O(nlogn)` over the whole refresh tokens table).