feat: add AtomicHandler to support live http.Handler reloads
This commit provides a basic primitive for the thread safe hot
swapping of an active http.Handler.

It works by wrapping an atomic.Value within a struct implementing
the http.Handler interface. Before serving each request it first
makes a load call on the atomic.Value, returning the most recently
stored http.Handler.

The package is accompanied by two tests, one to test the basic
store & load support. The second test is meant to be ran only when
the race detector is present.