feat: configurable email and sms rate limiting
Adds two new configuration values for rate limiting the sending
of emails and sms messages:

 - GOTRUE_RATE_LIMIT_EMAIL_SENT
 - GOTRUE_RATE_LIMIT_SMS_SENT

It is implemented with a simple rate limiter that resets a counter
at a regular interval. The first intervals start time is set when
the counter is initialized. It will be reset when the server is
restarted, but preserved when the config is reloaded.

Syntax examples:
  1.5       # Allow 1.5 events over 1 hour (legacy format)
  100       # Allow 100 events over 1 hour (1h is default)
  100/1h    # Allow 100 events over 1 hour (explicit duration)
  100/24h   # Allow 100 events over 24 hours
  100/72h   # Allow 100 events over 72 hours (use hours for days)
  10/30m    # Allow 10  events over 30 minutes
  3/10s     # Allow 3   events over 10 seconds

Syntax in ABNF to express the format as value:
  value = count / rate
  count = 1*DIGIT ["." 1*DIGIT]
  rate = 1*DIGIT "/" ival
  ival = ival-sec / ival-min / ival-hr
  ival-sec = 1*DIGIT "s"
  ival-min = 1*DIGIT "s"
  ival-hr = 1*DIGIT "h"

Co-authored-by: Stojan Dimitrovski <sdimitrovski@gmail.com>