feat: add support for LIFO ordering of after hooks
The goal here is to ensure `after-*` hooks were triggered as late
as possible in the request flow while maintaining intuitive
ordering. For example in the signup flow we want to ensure that the
`after-user-created` hook runs once all db triggers and identities
have been created for the user.

To support this we added the package `hookafter`. This package
stores a list of `after-*` triggers via `Defer` in the current
context to later be excuted a single time with `Fire`. A simple
middleware was added which returns context with the state present
and then queues a call to `Fire` in a defer before calling the
next handler. The call to `Fire` will run each previously queued
`Defer` in LIFO order.

This greatly simplifies triggering `after-*` hooks while
providing clear semantics for triggering multiple different
`after-*` hooks. For example:

A Signup flow calls into `signupNewUser`
- trigger: `before-user-created`
- a new user is added in `signupNewUser`
- defer: `after-user-created`
- the previous caller of `signupNewUser` adds a new identity
- defer: `after-identity-created`

No matter the caller of signupNewUser:
  (Invite, Mail, External or Signup, other?)

We ensure by calling `Defer` the ordering of hooks matches the
order of operations the flow followed.