feat: use pkg hookserrors in hookshttp & hookspgfunc
This allows consistent error handling across both the hookshttp
and hookspgfunc packages without the need to embed AuthHookError
in every single output struct.

In addition this required moving `validateTokenClaims` to the
API package. This removes the need for a single special case
in the dispatching of hooks in pkg `internal/hooks/v0hooks`.