chore: add send-email unit tests
First I added send email tests for the basic EmailChange flows
to better understand the existing behavior.

There is currently a bug in the existing send-email hooks which
omits the token during email changes for anonymous users when
the SecureEmailChange config value is set to false. I've also
identified a bug with the token hashes being mixed up. The tests
lock in this behavior so we don't break existing hooks code.