feat: ignore `aud` claim from admin jwt (`service_role` never had one) (#2070)
There's a problem with new Secret API keys which mint a JWT where the
`aud` claim is the requested resource. This is confusing list admin
users in returning no users (since there's no such audience).

`service_role` never had an `aud` claim in it, so this is the proper
place to fix this.