chore: make it more clear which rate limiters use the OTP limit (#2090)
In the supabase dashboard the rate limits are declared as:

- `RATE_LIMIT_OTP` Rate limit for sign ups and sign ins Number of sign
up and sign-in requests that can be made in a 5 minute interval per IP
address

- `RATE_LIMIT_VERIFY` Rate limit for token verifications Number of
OTP/Magic link verifications that can be made in a 5 minute interval per
IP address

This does not exactly translate to the rate limit usage in the auth
server. But it isn't something we may change, so for now I'm just making
this usage more clear and documenting it.

---------

Co-authored-by: Chris Stockton <chris.stockton@supabase.io>