feat: add request-scoped background task system
Introduce a new `apitask` package for running background tasks within the
lifecycle of an API request. We then use this feature in the mailer client to
send emails in the background when `GOTRUE_MAILER_EMAIL_BACKGROUND_SENDING`
is set to true. Before the request returns to the caller we first wait for all
in-flight tasks to exit. This is the smallest possible step I could come up with
to remove blocking email sending calls from the request path.

Summary:
* Define a `Task` interface for light background work.
* Add `Run(ctx, task)` to schedule tasks asynchronously when supported, or fall
  back to synchronous execution otherwise.
* Add `Wait(ctx)` to block until all request-scoped tasks finish
* Add `Middleware` to wrap HTTP handlers, attaching a request-scoped worker to
  the context and ensuring all tasks complete before returning a response.
* Add a check for `EmailBackgroundSending` in mailer package, when present
  wrap the mail client in a `backgroundMailClient`. When `Mail` is called it
  will send a `mailer.Task` to `apitasks.Run` to send the email in the
  background using the wrapped `MailClient`.