feat: introduce request-scoped background tasks & async mail sending (#2126)
## Summary
Introduce a new `apitask` package for running background tasks within
the lifecycle of an API request. We then use this feature in the mailer
client to send emails in the background. Before the request returns to
the caller we first wait for all in-flight tasks to exit.

**Summary:**
* Define a `Task` interface for light background work.
* Add `Run(ctx, task)` to schedule tasks asynchronously when supported,
or fall back to synchronous execution otherwise.
* Add `Wait(ctx)` to block until all request-scoped tasks finish
* Add `Middleware` to wrap HTTP handlers, attaching a request-scoped
worker to the context and ensuring all tasks complete before returning a
response.
* Add a check for `EmailBackgroundSending` in mailer package, when
present wrap the mail client in a `backgroundMailClient`. When `Mail` is
called it will send a `mailer.Task` to `apitasks.Run` to send the email
in the background using the wrapped `MailClient`.
* Add config `GOTRUE_MAILER_EMAIL_BACKGROUND_SENDING` (def `false`) -
on/off switch.

I will follow up with unit tests once the team has had an opportunity to
review.

---------

Co-authored-by: Chris Stockton <chris.stockton@supabase.io>