feat: implement link identity with oidc / native sign in (#2108)
Implements linking an OIDC identity to a user. It works by modifying the
existing `/token?grant_type=id_token` endpoint by allowing a
`link_identity` body parameter.

If it's set, the `Authorization` header must be set to a valid JWT
representing a user. This user will be the user to which the ID token's
identity will be linked to.