feat: rewrite all Nix expressions using blueprint framework
The flake now provides a straightforward devshell, a default
package (the binary built from the Go sources in this repository), and
an auth NixOS module with default settings that are turned into the
required configuration files and a systemd service. Some utility
scripts for managing the service have been removed, as we now expect
this to be handled directly through systemd utilities.