feat: notify users when their phone number has changed (#2184)
This PR adds support for sending the user an email notification when
their phone number has been changed.

3 new environment variables are introduced:

- `GOTRUE_MAILER_SUBJECTS_PHONE_CHANGED_NOTIFICATION`: Email subject to
use for phone changed notification.
- `GOTRUE_MAILER_TEMPLATES_PHONE_CHANGED_NOTIFICATION`: The URL to
specify a custom template.
- `GOTRUE_MAILER_NOTIFICATIONS_PHONE_CHANGED_ENABLED`: whether the
notification is enabled or not.

The feature is disabled by default. To enable it, the
`GOTRUE_MAILER_NOTIFICATIONS_PHONE_CHANGED_ENABLED` environment variable
must be set to `true`.