feat(oauth): add support for X/Twitter v2 provider (#2275)
Adds support for X (formerly Twitter) v2 as an external OAuth provider.

- Introduces an `oauth_client_states` table to persist the `code_verifier`s as
X mandates the use of PKCE
- Uses SHA256 challenge for PKCE
- Updates the `GetOAuthToken` signature to accept a `code_verifier` as
the second parameter
- Uses the existing cleanup middleware to delete states
- Adds the provider as `x` rather than `x_v2` as `twitter` is already
used in old OAuth 1.0a provider to better align with the rebrand
- The state is a UUIDv4

NOTE: today the `flow_states` table is overloaded, containing states,
auth codes, provider tokens...the goal is to decouple that table
eventually and the `oauth_states` table is the first step towards that.