chore: add unit tests for GenerateOtp (#2281)
This change locks in the current behavior of GenerateOtp with unit
tests. This required a small change to allow passing an io.Reader. In
addition I added a TODO comment about a minor defensive change to
include bounds checks, error propagation and the correct way to support
larger digits.

This is in response to https://github.com/supabase/auth/pull/2272

---------

Co-authored-by: Chris Stockton <chris.stockton@supabase.io>