Treat an empty first header value as a warning rather than an error
This commit updates performRateLimiting to emit a warning for headers
that start with a comma rather than return an error.

It is possible under some circumstances that Auth servers could
receive a rate limit header value of the form ", foo". Rather than
break functionality for servers using this, it is safer to instead
emit a warning and then change this to be an error in a future
release.