fix: check each type independently
Because type creation resides in the do block it is all or
nothing. When it's not a duplicate object that behavior is
appropriate, but if it's because one of the types already
exists then you will fail the creation of the other types.

For example:
* `aal_level` type already exists
* `factor_status` & `factor_type` are missing

Currently this would prevent `factor_status` & `factor_type` from
ever getting created.