chore: update Go from v1.23.7 to v1.25.5
As part of this upgrade several changes were needed to resolve go vet
failures around non-constant format strings:

- Updating all apierrors constructors to use const fmt strings with args
- Removing fmt.Sprintf usages that violate go vet fmt checks
- Refactoring internal error/message helpers to accept fmt + params

In addition stricter checks in the standard library for x509 certificate
creation required a change to a SAML test in internal/conf. Now I start with
a valid certificate and then set the serial number to an invalid value. To do
this I opted for a small refactor to PopulateFields to return the cert object
instead of copying the code within the test.